Added replace program

+++ stra.c
@@ -59,11 +59,11 @@ int Str_compare(const char s1[], const char s2[]) {
   i = 0;
   while (s1[i] != '\0' && s2[i] != '\0') {
     if (s1[i] != s2[i]) {
-      return s1[i] - s2[i];
+      return (int) s1[i] - (int) s2[i];
     }
     i++;
   }
-  return s1[i] - s2[i];
+  return (int) s1[i] - (int) s2[i];
 }
 
 char *Str_search(const char pcHaystack[], const char pcNeedle[]) {
