+++ strp.c
@@ -64,13 +64,13 @@ int Str_compare(const char *s1, const char *s2) {
 
     while (*s1 != '\0' && *s2 != '\0') {
         if (*s1 != *s2) {
-            return *s1 - *s2;
+            return (int) *s1 - (int) *s2;
         }
         s1++;
         s2++;
     }
 
-    return *s1 - *s2;
+    return (int) *s1 - (int) *s2;
 }
 
 char *Str_search(const char *pcHaystack, const char *pcNeedle) {
